[Core] Move FeatureSwitchService API to its own namespace
